Revert "Merge tag 'v2.6.31.2' of git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-2.6-stable"
This reverts commit 701ffc99afd34e46f1c0c2e2dc40fae79595cb03, reversing
changes made to edb5461b617ec4ba475393c11320140f41963de5.

+++ net/ax25/af_ax25.c
@@ -538,7 +538,7 @@ static int ax25_setsockopt(struct socket *sock, int level, int optname,
 	if (level != SOL_AX25)
 		return -ENOPROTOOPT;
 
-	if (optlen < (int)sizeof(int))
+	if (optlen < sizeof(int))
 		return -EINVAL;
 
 	if (get_user(opt, (int __user *)optval))
